refactor: rename mycelium.sstp → mycelium.protocol
The module holds CLI-side Pydantic models for two concerns: the SSTP wire
format (ProposeReply, RespondReply, NegotiatePayload, InboundTick) AND
the structured-memory category enum + MemoryLogEntry. The latter has
nothing to do with SSTP, so "sstp" was misleading. "protocol" honestly
covers both as "the CLI's shared wire-shape models."

- File renamed via git mv (history preserved)
- Imports updated in commands/memory.py and commands/negotiate.py
- Module-pointer comments updated in async_coordination.py and the
  structured-memory guide
- docs/index.html regenerated to pick up the rename
